+++ guide/homepage/.js/src/main/scala/io/udash/web/homepage/states.scala
@@ -2,8 +2,7 @@ package io.udash.web.homepage
 
 import com.avsystem.commons.misc.AbstractSealedEnumCompanion
 import io.udash._
-import io.udash.web.homepage.components.demo.DemoComponent
-import io.udash.web.homepage.components.demo.DemoComponent.CodeDemo
+import io.udash.web.homepage.components.demo.{CodeDemo, HelloDemo, SelectDemo}
 
 sealed abstract class RoutingState(val parentState: Option[ContainerRoutingState]) extends State {
   override type HierarchyRoot = RoutingState
@@ -16,9 +15,8 @@ case object RootState extends ContainerRoutingState(None)
 case object ErrorState extends RoutingState(Some(RootState))
 
 sealed abstract class IndexState(val name: String, val codeDemo: CodeDemo) extends RoutingState(Some(RootState))
-case object HelloState extends IndexState("Hello, World!", DemoComponent.HelloDemo)
-case object OtherState extends IndexState("Other", DemoComponent.HelloDemo)
+case object HelloState extends IndexState("Hello, World!", HelloDemo)
+case object SelectState extends IndexState("Select", SelectDemo)
 object IndexState extends AbstractSealedEnumCompanion[IndexState] {
   override val values: Seq[IndexState] = caseObjects
 }
-
